스크랩 / 참여 / 팔로우 기능 리서치 / 파트너 컨트롤러로 이관
