+++ src/Service/Partner/partner.find.service.ts
@@ -1,18 +1,6 @@
-import { Injectable, Inject } from "@nestjs/common";
-import { MongoPartnerFindService } from "src/Mongo";
+import { Injectable } from "@nestjs/common";
 
 @Injectable()
 export class PartnerFindService {
   constructor() {}
-
-  @Inject()
-  private readonly mongoPartnerFindService: MongoPartnerFindService;
-
-  async testPartnerRouter() {
-    return "testPartnerRouter@partner.find.service";
-  }
-
-  async getPartners() {
-    return await this.mongoPartnerFindService.getPartners();
-  }
 }
